moved tests to tests.py
